Refactor initial_setup.sh

+++ initial_setup.sh
@@ -1,23 +1,27 @@
 #!/usr/bin/env sh
 
+ansible_path=~/Library/Python/3.9/bin/ansible-pull
+
 echo 'Updating system ...'
-softwareupdate --install --all --restart &> /dev/null
+sudo softwareupdate --install --all --restart >/dev/null 2>&1
 echo 'DONE'
 
 if ! (xcode-select -p); then
     echo 'Installing Command Line Tools (you can track progress in System Settings > General > Software Update)...'
-    xcode-select --install &> /dev/null
+    xcode-select --install >/dev/null 2>&1
     softwareupdate --install --all
     echo ' DONE'
 fi
 
-echo 'Upgading Python3 ...'
-sudo pip3 install --upgrade pip
-echo ' DONE'
+if ! [ -e $ansible_path ]; then
+    echo 'Upgrading Python3 ...'
+    sudo pip3 install --upgrade pip
+    echo ' DONE'
 
-echo 'Installing Ansible ...'
-pip3 install ansible
-echo ' DONE'
+    echo 'Installing Ansible ...'
+    pip3 install ansible
+    echo ' DONE'
+fi
 
-echo -n 'Provisioning using `ansible-pull`'
-~/Library/Python/3.9/bin/ansible-pull -U https://github.com/dluksza/mac-dev-setup.git
+echo 'Provisioning using ansible-pull'
+$ansible_path -U https://github.com/dluksza/mac-dev-setup.git
